Add files via upload
A projects that aims to create a Wine Quality Analytics System to help determine the quality of the wines produced based on their composition. The project uses different supervised machine learning models that solve a classification problem. The project  focuses on Python and Statistics core concepts and its applications in Machine Learning.